julep-workflow/requirements.txt
julep
requests
python-dotenv
pyyaml
streamlit